Move location-options to under10k directory
This location option logic is only used on the simple form now so should
live there to discourage reuse from other forms.
